AB#244944
Update request body to comply with OpenAPI spec

Refactor the request body construction in the FileShareReadWriteClient to create an array of objects with `filename` and `hash` properties from the `FileDetails` of a `BatchHandle`. If `FileDetails` is null, an empty list is initialized. The `HttpRequestMessage` now uses this new request body instead of the previous `batchHandle`.