Add [FromBody] attribute to POST parameters
Updated the `S100CustomExchangeSetApiRouteBuilderExtension.cs` file to include the `[FromBody]` attribute for the `productNamesRequest` and `productVersionsRequest` parameters in two POST endpoint methods. This change clarifies that these parameters should be deserialized from the request body. Additionally, a new using directive for `Microsoft.AspNetCore.Mvc` was added to support this functionality.