247427 set expiry previous batches (#169)
* Filter Current Batch id

* commented  deployment stages for testing AB#247427

* updated the pipeline AB#247427

* Extend batch expiry date by two days

Updated the expiry date for batches in `DefaultFileService.cs` to be two days from the current UTC date, enhancing the validity period of the batches.

* Update batch expiry date to current time

Modified the `SetExpiryDateAsync` method to set the expiry date for a batch to the current UTC time instead of two days in the future. This change removes the previous two-day delay in setting the expiry date.

* Configuration change for testing

* change configuration for testing

* Update ExchangeSetGenerationSchedule to daily at midnight

Changed the Quartz Cron expression for `ExchangeSetGenerationSchedule` in `configuration.json` from `"0 * * * * ?"` to `"0 0 0 * * ?"`, scheduling the ExchangeSet generation to occur daily at midnight instead of every minute.

* Enhance DefaultFileService with new query filters

Updated the `DefaultFileService` class to include new query clauses: `ExpiryDateQueryClause` and `MediaTypeQueryClause`, allowing for more specific filtering of committed batches. The `SearchCommittedBatchesExcludingCurrentAsync` method now utilizes these clauses in its filter string, checking for `BusinessUnit`, `ProductCode`, `ExpiryDate`, and `MediaType`. Additionally, improved the readability of entry filtering by using the spread operator (`[..]`).

* Comment out dependencies for DevDeploy stage in azure-pipelines.yml

* Uncomment VulnerabilityChecks and Test stages in azure-pipelines.yml; update DevDeploy dependencies

---------

Co-authored-by: Tejasi Chavan <tejasi15072@mastek.com>
Co-authored-by: Suraj Singh <suraj49470@gmail.com>