Fixed Review comment AB#249075
Add BusinessUnit configuration for improved flexibility

This commit introduces a "BusinessUnit" key in the
`configuration.json` file across multiple sections
(orchestrator, dev, vni, vne, iat), setting it to
"ADDS-S100".

In `DefaultFileService.cs`, the hardcoded BusinessUnit
string has been replaced with a reference to the new
configuration key `BusinessUnitConfigKey`. This change
ensures that the BusinessUnit value is dynamically
retrieved from the configuration, enhancing the
configurability of the application.