Refactor CreateFilesArray method parameters
Updated the `CreateFilesArray` method in `FssResponseGenerator.cs` to change the `updateNo` parameter from a string to an int, renaming it to `updateNumber`. Removed the string parsing logic and updated method calls accordingly. Modified the handling of `expiryDate` to use `DateTime.UtcNow.AddMonths(2)` instead of null, and ensured consistent default value handling by passing `0` for `updateNumber` in one instance.